金額表示の32bit上限問題

+++ src/main/java/red/man10/MarketData.java
@@ -1906,7 +1906,7 @@ public class MarketData {
 
 
     public String getPriceString(double price){
-        return String.format("%,d",(int)(price));
+        return String.format("%,d",(long)(price));
     }
     //1,301 などを1301.0に変換する
     public Double getPricedouble(String price){
